Restyled by black

+++ bot.py
@@ -21,21 +21,26 @@ from get_enviroment import SECURITY_CHANNEL
 from get_enviroment import SECURITY_GUILD
 from get_enviroment import SWEAR_WORDS_LIST
 from get_enviroment import TOKEN
+
 # database import & connection
 
 conn = sqlite3.connect("maindatabase1.db")
 cur = conn.cursor()
-cur.execute("""CREATE TABLE IF NOT EXISTS `warns` (
+cur.execute(
+    """CREATE TABLE IF NOT EXISTS `warns` (
         `id` INTEGER PRIMARY KEY AUTOINCREMENT,
         `userid` INT(100),
         `guildid` INT,
         `reason` TEXT,
         `timestamp` INT);
-        """)
-cur.execute("""CREATE TABLE IF NOT EXISTS `settings` (
+        """
+)
+cur.execute(
+    """CREATE TABLE IF NOT EXISTS `settings` (
         `guildid` INT(100) UNIQUE,
         `automod` INT);
-        """)
+        """
+)
 
 hammericon = "https://images-ext-2.discordapp.net/external/OKc8xu6AILGNFY3nSTt7wGbg-Mi1iQZonoLTFg85o-E/%3Fsize%3D1024/https/cdn.discordapp.com/avatars/591633652493058068/e6011129c5169b29ed05a6dc873175cb.png?width=670&height=670"
 
@@ -53,9 +58,9 @@ bot.remove_command("help")
 #
 
 
-@bot.slash_command(name="help",
-                   description="Displays all the available commands for Hammer"
-                   )
+@bot.slash_command(
+    name="help", description="Displays all the available commands for Hammer"
+)
 async def help(ctx):
     # Define each page
 
@@ -64,9 +69,9 @@ async def help(ctx):
     **Hammer's commands:**
     """
 
-    embed = Embed(title="Hammer Bot Help",
-                  description=descr,
-                  colour=discord.Colour.lighter_grey())
+    embed = Embed(
+        title="Hammer Bot Help", description=descr, colour=discord.Colour.lighter_grey()
+    )
 
     embed.add_field(
         name="Moderation Commands :tools:",
@@ -125,8 +130,9 @@ async def help(ctx):
         inline=True,
     )
 
-    embed.set_footer(text=f"Hammer | Command executed by {ctx.author}",
-                     icon_url=hammericon)
+    embed.set_footer(
+        text=f"Hammer | Command executed by {ctx.author}", icon_url=hammericon
+    )
 
     await ctx.respond(embed=embed)
 
@@ -142,7 +148,7 @@ async def respondNotifOwner(text):
 
 
 async def GetWarnings(userid: int, fullData: bool = False):
-    cur.execute("SELECT * FROM warns WHERE userid=?", (userid, ))
+    cur.execute("SELECT * FROM warns WHERE userid=?", (userid,))
     rows = cur.fetchall()
     if not fullData:
         return len(rows)
@@ -183,8 +189,7 @@ async def Removewarn(userid: int, guildId: int, relativeWarnId: int):
 
 async def Clearwarns(userid: int, guildId: int):
     # delete all rows
-    cur.execute("DELETE FROM warns WHERE userid=? AND guildid=?",
-                (userid, guildId))
+    cur.execute("DELETE FROM warns WHERE userid=? AND guildid=?", (userid, guildId))
     conn.commit()
     return
 
@@ -207,8 +212,7 @@ async def getAllWarns(userid: int):
 
 
 async def GetSettings(guildid: int):
-    cur.execute("SELECT * FROM settings WHERE guildid = ? LIMIT 1",
-                (guildid, ))
+    cur.execute("SELECT * FROM settings WHERE guildid = ? LIMIT 1", (guildid,))
     rows = cur.fetchall()
     if len(rows) > 0:
         return rows[0][1]
@@ -217,13 +221,10 @@ async def GetSettings(guildid: int):
 
 
 async def SaveSetting(guildid: int, module: str, value: int):
-    cur.execute("SELECT * FROM settings WHERE guildid = ? LIMIT 1",
-                (guildid, ))
+    cur.execute("SELECT * FROM settings WHERE guildid = ? LIMIT 1", (guildid,))
     rows = cur.fetchall()
     # print(rows)
-    if len(
-            rows
-    ) > 0:  # cur.execute('INSERT INTO foo (a,b) values (?,?)', (strA, strB))
+    if len(rows) > 0:  # cur.execute('INSERT INTO foo (a,b) values (?,?)', (strA, strB))
         query = f"""UPDATE settings
         SET automod = {value}
         WHERE guildid={guildid} """
@@ -329,15 +330,17 @@ async def on_message(message):
             descr = (
                 f"The user {member} has been warned because said a banned swear word"
             )
-            embed = Embed(title=f"{member} has been warned! :hammer_pick:",
-                          description=descr)
+            embed = Embed(
+                title=f"{member} has been warned! :hammer_pick:", description=descr
+            )
             embed.set_footer(
                 text=f"Hammer | Automod service",
                 icon_url=hammericon,
             )
             embed.set_thumbnail(url=member.display_avatar)
-            warn = await AddWarning(member.id, message.guild.id,
-                                    "Said a banned swear word")
+            warn = await AddWarning(
+                member.id, message.guild.id, "Said a banned swear word"
+            )
             s = "s" if warn > 1 else ""
             embed.add_field(
                 name="Warn count",
@@ -345,9 +348,12 @@ async def on_message(message):
                 inline=True,
             )
             bannedmessage = (
-                message.content[:message.content.find(word)] + "~~" + word +
-                "~~" +
-                message.content[message.content.find(word) + len(word):])
+                message.content[: message.content.find(word)]
+                + "~~"
+                + word
+                + "~~"
+                + message.content[message.content.find(word) + len(word) :]
+            )
             embed.add_field(
                 name="Message Removed:",
                 value=f"The removed message was \n||{bannedmessage}||",
@@ -363,15 +369,17 @@ async def on_message(message):
                 embed = ErrorEmbed(
                     await message.channel.send(
                         f"Could not deliver the message to the user {member}\n This may be caused because the user is a bot, has blocked me or has the DMs turned off. \n\n**But the user is warned** and I have saved it into my beautiful unforgettable database"
-                    ), )
+                    ),
+                )
     # if(str(message.content).startswith(COMMAND_PREFIX)):
     # print("command executed", message.content)
 
 
 @bot.event
 async def on_ready():
-    await bot.change_presence(activity=discord.Activity(
-        type=discord.ActivityType.watching, name="you"))
+    await bot.change_presence(
+        activity=discord.Activity(type=discord.ActivityType.watching, name="you")
+    )
     print("HAMMER BOT Ready!", datetime.datetime.now())
     botname = await bot.application_info()
     print("logged in as:", botname.name)
@@ -388,9 +396,7 @@ async def on_ready():
 debug = False
 
 
-@bot.slash_command(guild_only=True,
-                   name="hello",
-                   guild_ids=[int(SECURITY_GUILD)])
+@bot.slash_command(guild_only=True, name="hello", guild_ids=[int(SECURITY_GUILD)])
 async def hello(ctx):
     await ctx.respond("Hammer is back!")
 
@@ -459,7 +465,8 @@ async def whois(ctx, member: discord.Member):
     description="Keeps out a user permanently, forbidding its entry",
 )
 @discord.default_permissions(
-    ban_members=True, )
+    ban_members=True,
+)
 async def ban(ctx, member: discord.Member, *, reason=None):
 
     if member == ctx.author:
@@ -470,8 +477,7 @@ async def ban(ctx, member: discord.Member, *, reason=None):
     message = f"You have been banned from {ctx.guild.name} for {reason}"
 
     descr = f"The user {member} has been banned for {reason}"
-    embed = Embed(title=f"{member} has been banned! :hammer_pick:",
-                  description=descr)
+    embed = Embed(title=f"{member} has been banned! :hammer_pick:", description=descr)
     embed.set_image(url="https://i.imgflip.com/19zat3.jpg")
     embed.set_footer(
         text=f"Hammer | Command executed by {ctx.author}",
@@ -494,11 +500,12 @@ async def ban(ctx, member: discord.Member, *, reason=None):
     await SendMessageTo(ctx, member, message)
 
 
-@bot.slash_command(guild_only=True,
-                   name="kick",
-                   description="Kicks out a member from the server")
+@bot.slash_command(
+    guild_only=True, name="kick", description="Kicks out a member from the server"
+)
 @discord.default_permissions(
-    kick_members=True, )
+    kick_members=True,
+)
 async def kick(ctx, member: discord.Member, *, reason=None):
 
     if member == ctx.author:
@@ -519,8 +526,7 @@ async def kick(ctx, member: discord.Member, *, reason=None):
             )
             return
     descr = f"The user {member} has been kicked for {reason}"
-    embed = Embed(title=f"{member} has been kicked! :hammer_pick:",
-                  description=descr)
+    embed = Embed(title=f"{member} has been kicked! :hammer_pick:", description=descr)
     embed.set_footer(
         text=f"Hammer | Command executed by {ctx.author}",
         icon_url=hammericon,
@@ -537,7 +543,8 @@ async def kick(ctx, member: discord.Member, *, reason=None):
     description="Sets a warning for a user, at 3 warns/strikes they get kicked",
 )
 @discord.default_permissions(
-    administrator=True, )
+    administrator=True,
+)
 async def warn(ctx, member: discord.Member, reason=None):
 
     if member == ctx.author:
@@ -548,8 +555,7 @@ async def warn(ctx, member: discord.Member, reason=None):
     message = f"You have been warned for {reason}"
 
     descr = f"The user {member} has been warned for {reason}"
-    embed = Embed(title=f"{member} has been warned! :hammer_pick:",
-                  description=descr)
+    embed = Embed(title=f"{member} has been warned! :hammer_pick:", description=descr)
     embed.set_footer(
         text=f"Hammer | Command executed by {ctx.author}",
         icon_url=hammericon,
@@ -573,7 +579,8 @@ async def warn(ctx, member: discord.Member, reason=None):
     description="Displays the warn history of a user in the guild",
 )
 @discord.default_permissions(
-    administrator=True, )
+    administrator=True,
+)
 async def seewarns(ctx, member: discord.Member):
     allwarns = await getAllWarns(member.id)
     if len(allwarns) == 0:
@@ -587,11 +594,12 @@ async def seewarns(ctx, member: discord.Member):
     return await ctx.respond(embed=embed)
 
 
-@bot.slash_command(guild_only=True,
-                   name="unwarn",
-                   description="Removes a strike from a user")
+@bot.slash_command(
+    guild_only=True, name="unwarn", description="Removes a strike from a user"
+)
 @discord.default_permissions(
-    kick_members=True, )
+    kick_members=True,
+)
 async def unwarn(ctx, member: discord.Member, id: int = None, *, reason=None):
     if await GetWarnings(member.id) == 0:
         return await ctx.respond("This user does not have any warn!")
@@ -600,8 +608,9 @@ async def unwarn(ctx, member: discord.Member, id: int = None, *, reason=None):
             f"""To select a warn to remove, use argument id and specify its value."""
         )
 
-        embed = Embed(title=f"ERROR! Need to select a warn :hammer_pick:",
-                      description=message)
+        embed = Embed(
+            title=f"ERROR! Need to select a warn :hammer_pick:", description=message
+        )
         allwarns = await getAllWarns(member.id)
         embed.add_field(
             name=f"**Historic of {member.name}**:",
@@ -613,8 +622,7 @@ async def unwarn(ctx, member: discord.Member, id: int = None, *, reason=None):
     message = f"You have been unwarned for {reason}"
 
     descr = f"The user {member} has been unwarned for {reason}"
-    embed = Embed(title=f"{member} has been unwarned! :hammer_pick:",
-                  description=descr)
+    embed = Embed(title=f"{member} has been unwarned! :hammer_pick:", description=descr)
     embed.set_footer(
         text=f"Hammer | Command executed by {ctx.author}",
         icon_url=hammericon,
@@ -632,11 +640,12 @@ async def unwarn(ctx, member: discord.Member, id: int = None, *, reason=None):
     await SendMessageTo(ctx, member, message)
 
 
-@bot.slash_command(guild_only=True,
-                   name="clearwarns",
-                   description="Removes all strikes from a user")
+@bot.slash_command(
+    guild_only=True, name="clearwarns", description="Removes all strikes from a user"
+)
 @discord.default_permissions(
-    kick_members=True, )
+    kick_members=True,
+)
 async def clearwarns(ctx, member: discord.Member, *, reason=None):
 
     if reason == None:
@@ -752,15 +761,16 @@ async def restart(ctx):
     description="Updates the message delay in a channel with a set of custom time interval",
 )
 @discord.default_permissions(
-    manage_messages=True, )
+    manage_messages=True,
+)
 async def setdelay(ctx, seconds: float, reason: str = ""):
 
     m = "modified" if seconds > 0.0 else "removed"
     embed = Embed(
         title=f"Delay {m} on #{ctx.channel} :hammer_pick:",
         description=f"This channel now has a delay of **{seconds}** seconds for {reason}"
-        if reason != None and reason != "" else
-        f"This channel now has a delay of **{seconds}** seconds",
+        if reason != None and reason != ""
+        else f"This channel now has a delay of **{seconds}** seconds",
     )
     embed.set_footer(
         text=f"Hammer | Command executed by {ctx.author}",
@@ -778,7 +788,8 @@ async def setdelay(ctx, seconds: float, reason: str = ""):
     description="Removes the hability to talk or join voice channels to a user",
 )
 @discord.default_permissions(
-    manage_messages=True, )
+    manage_messages=True,
+)
 async def mute(ctx, member: discord.Member, *, reason=None):
 
     guild = ctx.guild
@@ -821,7 +832,8 @@ async def mute(ctx, member: discord.Member, *, reason=None):
     description="Restores the hability to talk or join voice channels to a user",
 )
 @discord.default_permissions(
-    manage_messages=True, )
+    manage_messages=True,
+)
 async def unmute(ctx, member: discord.Member, *, reason=None):
 
     mutedRole = discord.utils.get(ctx.guild.roles, name="Muted")
@@ -831,8 +843,8 @@ async def unmute(ctx, member: discord.Member, *, reason=None):
         reason = "for " + reason
     await member.remove_roles(mutedRole)
     SendMessageTo(
-        ctx, member,
-        f":tada: You have been unmuted from: {ctx.guild.name} {reason}")
+        ctx, member, f":tada: You have been unmuted from: {ctx.guild.name} {reason}"
+    )
     embed = discord.Embed(
         title=f"User Unmuted: {member}",
         description=f"User {member.mention} has been unmuted {reason}",
@@ -931,9 +943,9 @@ modules = ["automod"]
 
 
 @discord.default_permissions(administrator=True)
-@bot.slash_command(name="settings",
-                   description="Modifies some Hammer config values",
-                   guild_only=True)
+@bot.slash_command(
+    name="settings", description="Modifies some Hammer config values", guild_only=True
+)
 @option(
     "module",
     description="Pick a module to switch!",
@@ -950,12 +962,10 @@ async def settings(ctx, module: str = None, value: str = None):
             value = 1 if value == "on" else 0
             await SaveSetting(ctx.guild.id, module, value)
             action = "enabled" if value else "disabled"
-            await ctx.respond(f"Module {module} {action} successfully!",
-                              ephemeral=True)
+            await ctx.respond(f"Module {module} {action} successfully!", ephemeral=True)
             return
         else:
-            await ctx.respond("Use: ``/settings module on/off``",
-                              ephemeral=True)
+            await ctx.respond("Use: ``/settings module on/off``", ephemeral=True)
             return
     embed = Embed(
         title=f"Hammer Bot Settings :hammer_pick:",
@@ -966,8 +976,9 @@ async def settings(ctx, module: str = None, value: str = None):
     automodStatustr = "**✅ ON**" if automodStatus else "**❌ OFF**"
     recommendedactivityAutomod = (
         f"Disable it by doing: ``{COMMAND_PREFIX}settings automod off``"
-        if automodStatus else
-        f"Enable it by doing ``{COMMAND_PREFIX}settings automod on``")
+        if automodStatus
+        else f"Enable it by doing ``{COMMAND_PREFIX}settings automod on``"
+    )
     embed.add_field(
         name="AutoMod Services :robot:",
         value=f"Actual status: {automodStatustr}\n {recommendedactivityAutomod}",
